+++ libs/php-rrule-master/tests/i18n/xx.php
@@ -1,19 +1,5 @@
 <?php
 
-/**
- * Translation file for English language.
- *
- * Most strings can be an array, with a value as the key. The system will
- * pick the translation corresponding to the key. The key "else" will be picked
- * if no matching value is found. This is useful for plurals.
- *
- * Licensed under the MIT license.
- *
- * For the full copyright and license information, please view the LICENSE file.
- *
- * @author Rémi Lanvin <remi@cloudconnected.fr>
- * @link https://github.com/rlanvin/php-rrule
- */
 return array(
 	'yearly' => array(
 		'1' => 'yearly',